requirements.txt
streamlit==1.20.0
google-generativeai==1.0.0  # Adjust version as needed
